chore(lists): add related links to other services in search results page
